Add files via upload
The objective of this project is get an insight into the raw data provided to know the overall performance of the Pizza Sales, Client gave us few Metrics to brain storm and come up with a meaningful Dashboard.

The Parameters are as follows :-
A:- Total Revenue
B:- Average Order Value
C:- Total Pizza's Sold
D:- Total Order's
E:- Average Order per Pizza

Also from Analyst point of view we wanted to fetch the results from SQL and match it with the "Measures" created the Power BI using DAX formula. We were provided the ".csv" file.